* mfc-static: (23 commits)
   remove makeCopyCommon.cmd makeCopySources.cmd
  makeDist.cmd: remove call .\makeCopyCommon.cmd
   scripts/build_openssl.bat: statically link openssl libraries
   NetBox.vcxproj: update AdditionalDependencies
   remove DelayLoadDLLs
   DelayLoadDLLs
   <DelayLoadDLLs>shell32.dll</DelayLoadDLLs>
   remove #include <Shellapi.h> #include <shlobj.h>
   AdditionalDependencies
   README.md: update
   NetBox.vcxproj: Debug|x64
   makeDist.cmd
   NetBox.vcxproj: Release|x64
   CHERRY-PICK: fix x64 compilation
   NetBox.vcxproj: Release|Win32
   libtinyxml.vcxproj
  Atlsd.lib;windowscodecs.lib
   _AFXDLL
   _AFXDLL
   _AFXEXT _AFXDLL
  ...

+++ src/NetBox/FarDialog.cpp
@@ -1162,7 +1162,7 @@ nb::TRect TFarDialogItem::GetActualBounds()
                      GetDialogItem()->X2, GetDialogItem()->Y2);
 }
 //---------------------------------------------------------------------------
-unsigned int TFarDialogItem::GetFlags()
+size_t TFarDialogItem::GetFlags()
 {
     return GetDialogItem()->Flags;
 }
